Roll back uintptr_t cast changes after discussions with tedu, otto and others.
C11 6.5.6.9 says:
When two pointers are subtracted, both shall point to elements of the
same array object, or one past the last element of the array object; the
result is the difference of the subscripts of the two array elements.

In these cases the objects are arrays of char so the result is defined,
and we believe that the report is based on a compiler incorrectly trapping
on defined behaviour.

+++ src/lib/libc/string/strlcpy.c
@@ -18,7 +18,6 @@
 
 #include <sys/types.h>
 #include <string.h>
-#include <stdint.h>
 
 /*
  * Copy string src to buffer dst of size dsize.  At most dsize-1
@@ -47,11 +46,6 @@ strlcpy(char *dst, const char *src, size_t dsize)
 			;
 	}
 
-	/*
-	 * Cast pointers to unsigned type before calculation, to avoid signed
-	 * overflow when the string ends where the MSB has changed.
-	 * Return value does not include NUL.
-	 */
-	return((uintptr_t)src - (uintptr_t)osrc - 1); 
+	return(src - osrc - 1);	/* count does not include NUL */
 }
 DEF_WEAK(strlcpy);
